+++ src/components/CourseInfo/CourseInfo.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import Button from '../common/Button/Button';
+import { Button } from '../common';
 import { useParams, useNavigate } from 'react-router-dom';
 import {
 	formatCreationDate,
@@ -9,7 +9,8 @@ import {
 } from '../helpers';
 import { useAppSelector } from 'src/store/hooks';
 import { getAuthorsSelector, getCoursesSelector } from 'src/store/selectors';
-export const CourseInfo: FC = () => {
+
+const CourseInfo: FC = () => {
 	const authorsState = useAppSelector(getAuthorsSelector);
 	const coursesState = useAppSelector(getCoursesSelector);
 
@@ -21,6 +22,7 @@ export const CourseInfo: FC = () => {
 		: 'None';
 
 	const navigate = useNavigate();
+
 	return (
 		<>
 			<div className='flex flex-col justify-between h-full mt-16 pb-20 items-center basis-full gap-8 border'>
@@ -56,3 +58,4 @@ export const CourseInfo: FC = () => {
 		</>
 	);
 };
+export default CourseInfo;
