+++ src/components/Courses/components/SearchBar/SearchBar.tsx
@@ -1,11 +1,9 @@
 import React, { FC } from 'react';
-//import Input from 'src/components/common/Input/Input';
-import Button from 'src/components/common/Button/Button';
-import { useForm } from 'src/components/hooks/useForm';
 import * as Yup from 'yup';
-//import { CoursesContext } from 'src/context/CourseContext';
+import { Button } from 'src/components/common';
+import { useForm } from 'src/components/hooks/useForm';
 
-const SearchBar: FC = () => {
+export const SearchBar: FC = () => {
 	const validationSchema = Yup.object().shape({
 		query: Yup.string()
 			.min(2, 'Too Short!')
@@ -18,15 +16,11 @@ const SearchBar: FC = () => {
 		validationSchema
 	);
 
-	//const { queryCourses, setCourses, courses } = useContext(CoursesContext);
-
 	const handleSubmit = async (event: React.FormEvent) => {
 		event.preventDefault();
 		const isValid = await validateForm();
 
 		if (isValid) {
-			// Filter the courses based on the query
-			//queryCourses(query);
 			resetForm();
 		} else {
 			console.log(errors);
@@ -52,5 +46,3 @@ const SearchBar: FC = () => {
 		</form>
 	);
 };
-
-export default SearchBar;
