+++ src/components/Courses/components/index.ts
@@ -1,0 +1,2 @@
+export * from './CourseCard/CourseCard';
+export * from './SearchBar/SearchBar';
