+++ src/components/CreateCourse/components/index.ts
@@ -1,0 +1,3 @@
+export * from './AuthorItem/AuthorItem';
+export * from './Authors/Authors';
+export * from './Form';
