+++ src/components/common/Input/TextArea.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from 'react';
 import useErrorVisibility from 'src/components/hooks/useErrorVisibility';
 import { InputCommonProps } from 'src/types/common/types';
-// Textarea component props
 
+// Textarea component props
 type TextareaProps = Omit<
 	React.TextareaHTMLAttributes<HTMLTextAreaElement>,
 	'prefix'
