+++ src/components/common/index.ts
@@ -1,2 +1,5 @@
-export * from './Button/Button';
-export * from './Input/Input';
+export { default as TextArea } from './Input/TextArea';
+export { default as Button } from './Button/Button';
+export { default as Container } from './Container/Container';
+export { default as Input } from './Input/Input';
+export { default as FormTemplate } from './Templates/Form';
