+++ src/components/index.ts
@@ -1,0 +1,7 @@
+export { default as CourseInfo } from './CourseInfo/CourseInfo';
+export { default as Courses } from './Courses/Courses';
+export { default as CourseForm } from './CreateCourse/CourseForm';
+export { default as Header } from './Header/Header';
+export { default as Login } from './Login/Login';
+export { default as PrivateRoute } from './PrivateRoute/PrivateRoute';
+export { default as Registration } from './Registration/Registration';
