+++ src/index.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import './index.css';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import store from './store';
-
+import './index.css';
 const root = ReactDOM.createRoot(
 	document.getElementById('root') as HTMLElement
 );
